After Todo List

+++ resources/views/managers/todo.blade.php
@@ -14,36 +14,72 @@
 @include('includes.headers_sidebars')
 
 @section('content')
-	<div class="container">
-		<form id="todoForm">
-			<input type="hidden" id="id" name="id" value="">
-			<div class="form-group">
-				<label>Title</label>
-				<input id="title" type="text" name="title" class="form-control">
-			</div>
-			<div class="form-group">
-				<label>Description</label>
-				<input type="text" id="description" name="description" class="form-control"/>
-			</div>
-			<div class="form-group">
-				<label>Comments</label>
-				<textarea id="comments" name="comments" class="form-control">
-				</textarea>
-			</div>
-			<div class="form-group">
-				<input type="submit" value="Save" class="btn btn-primary">
-			</div>
-		</form>
-		<hr>
-		<ul id="todos" class="list-group">
-			
-		</ul>
+<div class="row">
+    <div class="col-md-12">
+        <div class="panel panel-default">
+        @include('includes.messages') 
+        	<div class="panel-heading vd_bg-grey">
+                <a id="toggle_btn" href="#" class="btn btn-success btn-sm">Add New Todo</a>
+            </div>
+            <div class="panel-body">
+				<div id="addtodo" style="display: none;" class="container">
+					<div class="col-md-6">
+						<form id="todoForm">
+							<input type="hidden" id="id" name="id" value="">
+							<div class="form-group">
+								<label>Title</label><span class="pull-right">ID:<span id="my_id"></span> </span>
+								<input id="title" type="text" name="title" class="form-control">
+							</div>
+							<div class="form-group">
+								<label>Description</label>
+								<input type="text" id="description" name="description" class="form-control"/>
+							</div>
+							<div class="form-group">
+								<label>Comments</label>
+								<textarea id="comments" name="comments" class="form-control">
+								</textarea>
+							</div>
+							<div class="form-group">
+								<input type="submit" value="Save" class="btn btn-primary">
+							</div>
+						</form>
+					</div>
+					<hr>
+				</div>
+				<div class="table-responsive">
+					<table class="table table-striped" id="data-tables">
+						<thead>
+							<tr>
+								<th>Id</th>
+								<th>Title</th>
+								<th>Description</th>
+								<th>Status</th>
+								<th>Priority</th>
+								<th>Comments</th>
+								<th>Created</th>
+								<th>Updated</th>
+							    <th>Action</th> 
+							</tr>
+						</thead>
+					</table>
+				</div> <!-- end table-responsive -->
+			</div> <!-- end panel-body -->
+		</div> <!-- end panel -->
 	</div>
+</div>
 @endsection
 
 @section('page_script')
+
 	<script type="text/javascript">
 		$(document).ready(function() {
+			$('#toggle_btn').click(function(){
+				$(this).text(function(i, text){
+				  $('#addtodo').toggle();
+		          return text === "Hide Add Todo Form" ? "Add New Todo" : "Hide Add Todo Form";
+		        })
+			    
+			 });
 
 			getTodos();
 			
@@ -58,29 +94,48 @@
 				if (id) {
 					updateTodo(id, title, description, comments);
 				} else {
-					addTodo(title, description, comments);
+					if(title=="") {
+						//do nothing
+					} else {
+						addTodo(title, description, comments);
+					}
 				}
+			});
+
+			//Delete event
+			$('body').on('click', '.viewLink', function(e){
+				let id = $(this).data('id');
+				$('#addtodo').show();
+				$('#toggle_btn').text('Hide Add Todo Form');
+				let mode = 'Viewing';
+				showTodo(id, mode);
 			});
 
 			//Delete event
 			$('body').on('click', '.deleteLink', function(e){
 				let id = $(this).data('id');
+				if (confirm('Are you sure you want to delete ID: '+ id)) {
+					deleteTodo(id);
+				}
 				
-				deleteTodo(id);
 			});
 
 			//update event
 			$('body').on('click', '.editLink', function(e){
 				let id = $(this).data('id');
-				
-				showTodo(id);
+				$('#addtodo').show();
+				$('#toggle_btn').text('Hide Add Todo Form');
+				let mode = 'Editing';
+				showTodo(id, mode);
 			});
 
 			//Show Todo using API
-			function showTodo(id) {
+			function showTodo(id, mode) {
 				$.ajax({
 					url: 'http://gofree.test/api/todos/'+id 
 				}).done(function(todo){
+					$('#title_hdr').html(mode+' ID: '+id);
+					$('#my_id').html(todo.id);
 					$('#id').val(todo.id);
 					$('#title').val(todo.title);
 					$('#description').val(todo.description);
@@ -100,10 +155,11 @@
 						comments: comments,
 					}
 				}).done(function(todo){
-					//alert('Todo # '+id+' was updated.');
-					$('#todos').html('');
-					getTodos();
-					clearForm();
+					alert('Todo # '+id+' was updated.');
+					location.reload();
+					//$('#todos').html('');
+					//getTodos();
+					//clearForm();
 				});
 			}
 			//clear the form
@@ -120,10 +176,10 @@
 					url: 'http://gofree.test/api/todos/'+id,
 					data: {_method: 'DELETE'}
 				}).done(function(todo){
-					//alert('Todo # '+id+' was deleted.');
-					//location.reload();
-					$('#todos').html('');
-					getTodos();
+					alert('Todo # '+id+' was deleted.');
+					location.reload();
+					//$('#todos').html('');
+					//getTodos();
 				});
 			}
 
@@ -138,40 +194,76 @@
 						comments: comments,
 					}
 				}).done(function(todo){
-					
-					//alert('Todo # '+todo.id+' was added.');
-					let output = '';
-					output += `
-						<li class="list-group-item">
-							<strong>TODO #${todo.id}: ${todo.title}: </strong>${todo.description} <a href="#" class="deleteLink" data-id="${todo.id}">Delete</a> <a href="#" class="editLink" data-id="${todo.id}">Edit</a>
-						</li>
-					`;
-
-					$('#todos').append(output);
-					clearForm();
+					location.reload();
 				});
 			}
 
 			//Get Todos from API
 			function getTodos() {
+				$('#data-tables').DataTable( {
+			        "ajax": "http://gofree.test/api/todos",
+			        "columns": [
+			            { "data": "id" },
+			            { "data": "title" },
+			            { "data": "description" },
+			            { "data": "status_id" },
+			            { "data": "priority" },
+			            { "data": "comments" },
+			            { "data": "created_at" },
+			            { "data": "updated_at" },
+			            { "data": "action" },
+
+			        ]
+			    } );
+			}	
+			
+			/*function getTodos2() {
 				$.ajax({
 					url: 'http://gofree.test/api/todos'
 				}).done(function(todos){
 					let output = '';
 					$.each(todos, function(key, todo){
 						output += `
+							<tr style="font-size: 12px;" class="py-0">
+								<td>${todo.id}</td>
+								<td>${todo.title}</td>
+								<td>${todo.description}</td>
+								<td>status</td>
+								<td>priority</td>
+								<td>target_date</td>
+								<td>Resource</td>
+								<td></td>
+								<td></td>
+								<td class="menu-action">
+									<a href="/admin/users/${todo.id}" data-original-title="view" data-toggle="tooltip" data-placement="top" class="btn menu-icon vd_bd-green vd_green"> <i class="fa fa-eye"></i> 
+									</a> 
+									<a href="/admin/users/${todo.id}/edit" data-original-title="edit" data-toggle="tooltip" data-placement="top" class="btn menu-icon vd_bd-yellow vd_yellow"> <i class="fa fa-pencil"></i> 
+									</a> 
+									<a href="/admin/users/${todo.id}" onclick="event.preventDefault(); document.getElementById('delete-form[]').submit();" data-original-title="delete" data-toggle="tooltip" data-placement="top" class="btn menu-icon vd_bd-red vd_red"> <i class="fa fa-times"></i> 
+									</a>
+								</td>
+							</tr>
+							
+						`;
+					});
+					$.each(todos, function(key, todo){
+						output += `
 							<li class="list-group-item">
-								<strong>TODO #${todo.id}: ${todo.title}: </strong>${todo.description} <a href="#" class="deleteLink" data-id="${todo.id}">Delete</a> <a href="#" class="editLink" data-id="${todo.id}">Edit</a>
+								<strong>TODO #${todo.id}: ${todo.title}: </strong>${todo.description} <a href="#" class="deleteLink btn btn-danger btn-sm" data-id="${todo.id}">Delete</a> <a href="#" class="editLink  btn btn-info btn-sm" data-id="${todo.id}">Edit</a>
 							</li>
 						`;
 					});
+				$('#todos').html('');
+				$('#todos').append(output);
+				$('#data-tables').dataTable().ajax.reload();
 
-					$('#todos').append(output);
+
 				});
-			}
+			} *///getTodo2
 
 		}) 
 
 	</script>
-    <?php //include(resource_path('views').'/templates/scripts/pages-user-profile.blade.php'); ?>
+
+	<?php include(resource_path('views').'/templates/scripts/listtables-data-tables.blade.php'); ?>
 @endsection
